requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
Werkzeug==0.10.1
WTForms==2.0.2
WTForms-Test==0.1.1
requests==2.5.3
unittest2==1.0.1
tweepy==3.3.0
python-instagram==1.3.0
six==1.9.0
simplejson==3.6.5
foursquare==20130707
wsgiref==0.1.2
traceback2==1.4.0
linecache2==1.0.0
httplib2==0.9
oauthlib==0.7.2
pip==1.5.6
poster==0.8.1
requests-oauthlib==0.4.2